requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2
openai==1.47.0
google-generativeai==0.8.3